Did refactoring and removed unwanted files
